+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Not released:
 
+* Fix `Konva.Transformer` behavior on mirrored nodes
 * Fix `stage.getPointerPosition()` logic.
 
 ## 4.0.8  - 2019-09-05